+++ app/src/main/java/de/hhn/softwarelabor/pawswipeapp/CreateUserProfileActivity.kt
@@ -18,6 +18,7 @@ import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import de.hhn.softwarelabor.pawswipeapp.api.user.ProfileApi
+import de.hhn.softwarelabor.pawswipeapp.utils.DatePickerFragment
 import java.io.ByteArrayOutputStream
 
 private const val PICK_IMAGE_REQUEST = 1
