+++ app/src/main/java/de/hhn/softwarelabor/pawswipeapp/FilterActivity.kt
@@ -11,6 +11,7 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.Spinner
 import androidx.appcompat.app.AppCompatActivity
+import de.hhn.softwarelabor.pawswipeapp.utils.DatePickerFragment
 
 class FilterActivity : AppCompatActivity() {
