+++ app/src/main/java/de/hhn/softwarelabor/pawswipeapp/SettingsActivity.kt
@@ -10,6 +10,7 @@ import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import de.hhn.softwarelabor.pawswipeapp.api.user.ProfileApi
+import de.hhn.softwarelabor.pawswipeapp.utils.DatePickerFragment
 
 
 class SettingsActivity : AppCompatActivity() {
